[#175] refactor :  기회비용 카드 ui를 alternatives 폴더로 이동
